@@ -1,9 +1,9 @@
-module github.com/shtirlic/postrint
+module github.com/shtirlic/posprint
 
-go 1.14
+go 1.21
 
 require (
 	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0
-	github.com/google/gousb v0.0.0-20190812193832-18f4c1d8a750
-	golang.org/x/image v0.5.0
+	github.com/google/gousb v1.1.2
+	golang.org/x/image v0.14.0
 )
